refactor: DTO 패키지 구조 변경
- `request` 및 `response` 패키지를 `dto` 하위에 추가하여 데이터 전송 객체의 구조를 개선
- `AuthResponse` 파일을 `dto/response` 패키지로 이동
